requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
neo4j==5.24.0
python-dotenv==1.0.1
pydantic>=2.0.0,<3.0.0
openai
dashscope
passlib
sqlalchemy
requests
urllib3
utils